+++ src/dependency_injector/__init__.py
@@ -1,6 +1,6 @@
 """Dependency injector top-level package."""
 
-__version__ = '3.13.1'
+__version__ = '3.13.2'
 """Version number that follows semantic versioning.
 
 :type: str